tests: split examples for better tinygo support.
