+++ internal/auditlog/auditlog_test.go
@@ -1,3 +1,6 @@
+// Copyright 2023 Juan Pablo Tosso and the OWASP Coraza contributors
+// SPDX-License-Identifier: Apache-2.0
+
 package auditlog
 
 import "testing"
